+++ contracts/modules/RangeEditionMinter.sol
@@ -178,8 +178,8 @@ contract RangeEditionMinter is IRangeEditionMinter, BaseMinter {
         uint32 quantity
     ) public view virtual override(BaseMinter, IMinterModule) returns (uint128) {
         unchecked {
-            // Won't overflow, as `price` is 96 bits, and `quantity` is 32 bits.
-            return _editionMintData[edition][mintId].price * quantity;
+            // Will not overflow, as `price` is 96 bits, and `quantity` is 32 bits. 96 + 32 = 128.
+            return uint128(_editionMintData[edition][mintId].price) * uint128(quantity);
         }
     }
